manage chatthreads internal to StatefulGraphChatListClient
